Introduce a server-side model class.
The model has the ability to invoke evaluation of each cell and publish
the results to the client, which displays them, though evaluation is
currently implemented only in the trivial case of state cells, not for
formulas.

The model also has API to write state cells and create and delete
columns with some limitations, which is used to load the sample data.

Next is to implement enough formula syntax to do simple joins and add
some kind of UI to the client to add and remove columns with such
formulas.