Make _unit a primitive type separate from the root column.
Down the road, we'll want to render the root cell ID consistently with
other cell IDs, and I expect it will confuse users to see references to
the root cell when all they want is a primitive unit type.

Also, since we aren't validating the value yet, render the actual value
so we can see problems.  And change the value to 'X' in
load-sample-data.