More compact, colored header.
A column's cellName cell is now hidden if it has no cellName and no
children.  Use the context menu to set the cellName.  Accordingly, the
column ID has been moved to the full text of the type cell.

Also add "not ready" checks for missing column types and families as
discussed.  In my tests, this seems to have completely stopped
intermediate states from appearing.