Delay re-highlighting so keyboard navigation isn't so slow.
Once we do this, we can afford to call onSelection from afterSelection,
which updates the action bar sooner.  The min-height of the action bar
was smaller than the formula bar in Matt's configuration, causing wild
oscillation as the appearance and disappearance of the formula bar
caused the corresponding cell to move out from under the mouse and back.
Increase it and change it back to a height so the content will be
clipped instead if it is too tall; if we see this happening, we can
increase the height further.  Obviously this will have to change if we
add back something like the "New Column" dialog (see 19cd313).