Update column name scoping rules.
A key column is now in the scope of its own object, not the parent
object.  This applies to column references (used in load-sample-data and
in the UI to specify types) as well as navigations.

When converting a column reference or navigation to a string, mark it
"(ambiguous)" or "(problem)" if re-parsing it would not give the
original meaning.

Finish support for using [fieldName] as an objectName for object types
without an objectName.

Remove the constraint that the fieldNames and objectNames of the
children of a given column be unique.  It isn't sufficient to ensure
that navigations are unambiguous, so I don't think it's worth continuing
to maintain.

Fix broken defineColumn call from the client from 1c72e3b.

Move definitions from src/common.coffee that are not specific to
Relational Spreadsheets to src/lib/util.coffee to make src/common.coffee
more manageable.