Remove the @ notation for references and add a default reference display format (pretty dumb at this point).
This is in response to overwhelming user feedback.  I would like to have
the @ notation available as an option, but we'll see if it's worth it.

Remove the caption column since we no longer need the ruler for the @
notation and I didn't think it was reasonable to reserve a whole column
for the "Obj" and "Field" labels.  I regret the loss of these labels,
but I thought it would be confusing to move them to the root object
column.  On the other hand, I did move the header expansion button to
the root object column.