Upgrade Handsontable to version 0.20.1.
I originally thought it might help me implement a dynamic context menu.
It didn't, and it was probably a mistake not to check this before
investing the work, but now that I've started building on this version,
I think the inconvenience of going back outweighs the risk of
regressions not being caught and affecting the user study.