Do not allow formulas to access tokens.
Previously, if you entered a formula referring to a state column and
then wrapped that column in objects, the formula would get type _token
and the UI would freak out.  Arguably we should change the "wrap"
command so that the original column ID goes to the field, not the
object, but this fix is valid regardless.