Check in TypeScript files.
We can now edit them manually, and the new "merge-from-coffeescript"
tool is able to merge changes to the automated conversion output with
manual edits.  Also check in a copy of the automated conversion output
for "merge-from-coffeescript".

fallback.ts is no longer a special case, but is treated as a manual edit
(currently the only one).