Get the TypeScript property accumulator working.
Unfortunately, the illegal let on the destructuring assignment in
src/language/procedures.ts (which I edited manually) was blocking tsc
from getting to the semantic analysis phase to issue the errors that the
property accumulator looks for.  Rather than introducing a way to patch
the TypeScript files in the middle of the conversion process, I just
made the converter smarter about destructuring assignments.