Switch to a TypeScript build plugin based on the Hansoft one.
This plugin places all files in one scope, so we won't have to use
<reference>.

The custom plugin is in a new submodule; be sure to initialize it.  I've
given the local package a distinct version number (1.1.1-objsheets) and
pinned the app to that version number so that if you neglect to
initialize the submodule, Meteor will fail instead of using the
published package (which is missing functionality we need).