Change Template.foo to Template['foo'] to remove TypeScript errors.
We need to get to zero errors first before we can start actually
enforcing types on different parts of the app (in whatever order we
like).  To actually check the templates, we'll probably have to use an
approach like the one at
http://blog.dataflows.io/technology/2015/04/29/meteor_typescript.html.

Bulk edit made on CoffeeScript and merged to TypeScript.  I prefer to
bulk edit CoffeeScript rather than TypeScript when it's similar in
difficulty to avoid increasing the number of merge conflicts we'll have
if we later make an overlapping change to the CoffeeScript conversion.