Fix variables that were declared with too small a scope by the insufficiently smart converter.
The TypeScript compiler leaves unresolvable names as-is, effectively
generating the same code as if the converter had used the original "var"
instead of doing the dumb replacement with "let".  However, TypeScript
1.8.2 (not yet enabled in the Meteor build plugin) rigorously clears and
renames "let" variables to enforce their scoping.  So if the generated
JavaScript looks equivalent before and after (which it does to me), it's
evidence that this commit correctly maintains the semantics of the
original CoffeeScript code that used function-scoped variables.