Split the apps into "client" and "server" subdirectories.
Commit 65f916a broke the Relsheets object into RelsheetsClient and
RelsheetsServer objects as the simplest way (albeit with some churn) to
have both the client and server declarations properly exported for VS
Code but have the Meteor client and server builds include only the
applicable code.  That commit resolved a bunch of TypeScript errors in
VS Code, but I realized the references from the app files to
RelsheetsClient and RelsheetsServer from "if Meteor.isClient" and "if
Meteor.isServer" statements were still generating TypeScript errors in
the Meteor builds.  This split is again the simplest way to solve the
problem, with some churn.

Edited in CoffeeScript and merged to TypeScript.

This change somehow exposed a latent race condition in the converter,
which is fixed.