Impose coding guidelines.
They are described in the new CODING_GUIDELINES.md file.  A major
component is setting up tslint (a TypeScript linter) with a set of lint
rules.

Correct existing violations of the more specific guidelines.  The most
common issues:

- "var" -> "let"
- Variable shadowing
- Document "any" or replace it with "fixmeAny"

Take the opportunity to update README.md and add DEVELOPMENT.md with
some notes about the development infrastructure.